+++ services/horizon/internal/actions/helpers_test.go
@@ -6,6 +6,7 @@ import (
 	"math"
 	"net/http"
 	"net/url"
+	"strings"
 	"testing"
 
 	"github.com/go-chi/chi"
@@ -141,7 +142,7 @@ func TestValidateCursorWithinHistory(t *testing.T) {
 			if tc.valid {
 				tt.NoError(err)
 			} else {
-				tt.EqualError(err, "problem: before_history")
+				tt.True(strings.Contains(err.Error(), "problem: before_history"))
 			}
 		})
 	}