+++ services/horizon/internal/init.go
@@ -5,13 +5,13 @@ import (
 	"net/http"
 	"runtime"
 
-	"github.com/stellar/go/services/horizon/internal/paths"
-
 	"github.com/getsentry/raven-go"
 	"github.com/prometheus/client_golang/prometheus"
+
 	"github.com/stellar/go/exp/orderbook"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/ingest"
+	"github.com/stellar/go/services/horizon/internal/paths"
 	"github.com/stellar/go/services/horizon/internal/simplepath"
 	"github.com/stellar/go/services/horizon/internal/txsub"
 	"github.com/stellar/go/services/horizon/internal/txsub/sequence"
@@ -102,22 +102,24 @@ func initIngester(app *App) {
 		HistorySession: mustNewDBSession(
 			db.IngestSubservice, app.config.DatabaseURL, ingest.MaxDBConnections, ingest.MaxDBConnections, app.prometheusRegistry,
 		),
-		NetworkPassphrase:            app.config.NetworkPassphrase,
-		HistoryArchiveURLs:           app.config.HistoryArchiveURLs,
-		CheckpointFrequency:          app.config.CheckpointFrequency,
-		StellarCoreURL:               app.config.StellarCoreURL,
-		StellarCoreCursor:            app.config.CursorName,
-		CaptiveCoreBinaryPath:        app.config.CaptiveCoreBinaryPath,
-		CaptiveCoreStoragePath:       app.config.CaptiveCoreStoragePath,
-		CaptiveCoreConfigUseDB:       app.config.CaptiveCoreConfigUseDB,
-		CaptiveCoreToml:              app.config.CaptiveCoreToml,
-		RemoteCaptiveCoreURL:         app.config.RemoteCaptiveCoreURL,
-		EnableCaptiveCore:            app.config.EnableCaptiveCoreIngestion,
-		DisableStateVerification:     app.config.IngestDisableStateVerification,
-		EnableReapLookupTables:       app.config.HistoryRetentionCount > 0,
-		EnableExtendedLogLedgerStats: app.config.IngestEnableExtendedLogLedgerStats,
-		RoundingSlippageFilter:       app.config.RoundingSlippageFilter,
-		EnableIngestionFiltering:     app.config.EnableIngestionFiltering,
+		NetworkPassphrase:                    app.config.NetworkPassphrase,
+		HistoryArchiveURLs:                   app.config.HistoryArchiveURLs,
+		CheckpointFrequency:                  app.config.CheckpointFrequency,
+		StellarCoreURL:                       app.config.StellarCoreURL,
+		StellarCoreCursor:                    app.config.CursorName,
+		CaptiveCoreBinaryPath:                app.config.CaptiveCoreBinaryPath,
+		CaptiveCoreStoragePath:               app.config.CaptiveCoreStoragePath,
+		CaptiveCoreConfigUseDB:               app.config.CaptiveCoreConfigUseDB,
+		CaptiveCoreToml:                      app.config.CaptiveCoreToml,
+		RemoteCaptiveCoreURL:                 app.config.RemoteCaptiveCoreURL,
+		EnableCaptiveCore:                    app.config.EnableCaptiveCoreIngestion,
+		DisableStateVerification:             app.config.IngestDisableStateVerification,
+		StateVerificationCheckpointFrequency: uint32(app.config.IngestStateVerificationCheckpointFrequency),
+		StateVerificationTimeout:             app.config.IngestStateVerificationTimeout,
+		EnableReapLookupTables:               app.config.HistoryRetentionCount > 0,
+		EnableExtendedLogLedgerStats:         app.config.IngestEnableExtendedLogLedgerStats,
+		RoundingSlippageFilter:               app.config.RoundingSlippageFilter,
+		EnableIngestionFiltering:             app.config.EnableIngestionFiltering,
 	})
 
 	if err != nil {